api/requirements.txt
falcon
gunicorn
feedparser
readability-lxml
requests
vaderSentiment
textblob